Broke all tests into own modules
